refactor: totalprice to int

+++ lib/providers/cart_provider.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:jajanan_boeng/models/cart_model.dart';
 import 'package:jajanan_boeng/models/product_model.dart';
-import 'package:jajanan_boeng/services/product_service.dart';
 
 class CartProvider with ChangeNotifier {
   List<CartModel> _carts = [];
@@ -60,7 +59,7 @@ class CartProvider with ChangeNotifier {
   }
 
   totalPrice() {
-    double total = 0;
+    int total = 0;
 
     for (var item in _carts) {
       total += (item.quantity * item.product.price);
